+++ tools/docker-entrypoint.sh
@@ -1,14 +1,22 @@
 #!/bin/bash
 
+if [ -z "$SETUP_POSTGRES" ] ; then
+    SETUP_POSTGRES="true"
+fi
+
+if [ -z "$PGVER" ] ; then
+    echo "[setup] PGVER is not set, using default version"
+    PGVER="17"
+fi
+
 if [ "$USE_PGLAB" = "true" ] ; then
     PGPATH=/pg_lab
 else
     PGPATH=/postbound/db-support/postgres
 fi
 
-if [ -z "$PGVER" ] ; then
-    echo "[setup] PGVER is not set, using default version"
-    PGVER="17"
+if [ -z "$SETUP_DUCKDB" ] ; then
+    SETUP_DUCKDB="false"
 fi
 
 if [ -z "$(ls /postbound)" ] ; then
@@ -44,17 +52,22 @@ if [ -z "$(ls /postbound)" ] ; then
     fi
 
     # Setup databases
-    cd /postbound/db-support/postgres
-    if [ "$SETUP_IMDB" = "true" ] ; then
-        ./workload-job-setup.sh --cleanup
+    if [ "$SETUP_POSTGRES" = "true" -a "$SETUP_IMDB" = "true" ] ; then
+        echo "[setup] Setting up IMDB benchmark for Postgres"
+        cd /postbound/db-support/postgres
+        ./workload-job-setup.sh
         ./postgres-psycopg-setup.sh job imdb
     fi
-    if [ "$SETUP_STATS" = "true" ] ; then
-        ./workload-stats-setup.sh --cleanup
+    if [ "$SETUP_POSTGRES" = "true" -a "$SETUP_STATS" = "true" ] ; then
+        echo "[setup] Setting up Stats benchmark for Postgres"
+        cd /postbound/db-support/postgres
+        ./workload-stats-setup.sh
         ./postgres-psycopg-setup.sh stats stats
     fi
-    if [ "$SETUP_STACK" = "true" ] ; then
-        ./workload-stack-setup.sh --cleanup
+    if [ "$SETUP_POSTGRES" = "true" -a "$SETUP_STACK" = "true" ] ; then
+        echo "[setup] Setting up Stack benchmark for Postgres"
+        cd /postbound/db-support/postgres
+        ./workload-stack-setup.sh
         ./postgres-psycopg-setup.sh stack stack
 
         cd /postbound/workloads/Stack-Queries
@@ -69,14 +82,50 @@ if [ -z "$(ls /postbound)" ] ; then
 
     # Install PostBOUND package
     cd /postbound && tools/setup-py-venv.sh --venv /postbound/pb-venv --skip-doc
+    source /postbound/pb-venv/bin/activate
+
+    # Setup DuckDB
+    # This needs to happen _after_ the PostBOUND package is installed, because we install the DuckDB Python package
+    # in the same virtual environment
+    if [ "$SETUP_DUCKDB" = "true" ] ; then
+        echo "[setup] Setting up DuckDB"
+        cd /postbound/db-support/duckdb
+        ./duckdb-setup.sh
+        . ./duckdb-env.sh
+    fi
+
+    if [ "$SETUP_DUCKDB" = "true" -a "$SETUP_IMDB" = "true" ] ; then
+        echo "[setup] Setting up IMDB benchmark for DuckDB"
+        cd /postbound/db-support/duckdb
+        ./workload-job-setup.sh
+    fi
+    if [ "$SETUP_DUCKDB" = "true" -a "$SETUP_STATS" = "true" ] ; then
+        echo "[setup] Setting up Stats benchmark for DuckDB"
+        cd /postbound/db-support/duckdb
+        ./workload-stats-setup.sh
+    fi
+    mv /postbound/db-support/duckdb/*.duckdb /postbound
+
+    # Clean up raw database files
+    echo "[setup] Cleaning up"
+    rm -rf /postbound/db-support/imdb_data
+    rm -rf /postbound/db-support/stats_data
+    rm -rf /postbound/db-support/stack_data
 
     # User config
-    if [ "$USE_PGLAB" = "true" ] ; then
+    if [ "$SETUP_POSTGRES" = "true" -a "$USE_PGLAB" = "true" ] ; then
         echo "cd /pg_lab/" >> /home/$USERNAME/.bashrc
-    else
+        echo "source ./postgres-load-env.sh" >> /home/$USERNAME/.bashrc
+    elif [ "$SETUP_POSTGRES" = "true" ] ; then
         echo "cd /postbound/db-support/postgres/" >> /home/$USERNAME/.bashrc
+        echo "source ./postgres-load-env.sh" >> /home/$USERNAME/.bashrc
     fi
-    echo "source ./postgres-load-env.sh" >> /home/$USERNAME/.bashrc
+
+    if [ "$SETUP_DUCKDB" = "true" ] ; then
+        echo "cd /postbound/db-support/duckdb" >> /home/$USERNAME/.bashrc
+        echo "source ./duckdb-env.sh" >> /home/$USERNAME/.bashrc
+    fi
+
     echo "cd /postbound" >> /home/$USERNAME/.bashrc
     echo "source /postbound/pb-venv/bin/activate" >> /home/$USERNAME/.bashrc